actually this stuping fucking keyboard is unusable without retro tapping
who the fuck thinks it's ok to drop the key combintation if it's pressed
and released before the tapping timeout? How fucking slow is everybody
typing? and the fucking code is unreadable, i'll have to spend a week
trying to figure out the places i'll have to change to make this stupid
fucking thing working properly, or I'll have to leave with retro typing?
jesus fucking christ please kill me